Start/Stop Kotlin functionality
When the start recording button is pressed the system will automatically record for 5 seconds and then stop recording and save the recorded file to the phones cache. The recording will immediately restart when it stops. File names will auto increment so no files are overridden.  The 5 seconds was used for testing and can be increased when needed.