Added expiry time for token and check for duplicate team numbers
Added an expiry time for the jwt created on login. Token lasts 24 hours, after which it will no longer work and a new one has to be created.

Added a check when creating a player to see if the number given to that player already exists for the team it is being added it, if it does exist the player will not be created.